Refactoring of redux
